+++ NEWS.md
@@ -2,6 +2,7 @@
 
 - Bug fix: `auto_scroll` now works better and smoother.
 - Bug fix: the enter key will not trigger page turn if working on `text`, `textarea`, `numeric`, `select`, `date`, and `daterange`.
+- Enhance: the Enter key can now used to trigger the "Submit and Exit" button for the rating pop-up modal.
 
 # surveydown 0.3.5